Moved are_same -> is_same
